refactor(api)!: widen block state ids from short to int
Minecraft 26.3 has 35723 block states, so the highest state id is 35722 and
no longer fits in a short (max 32767). 26.1 and 26.2 peaked at 29872 and
32365, which is why the short typing went unnoticed until now.

`SimpleBlock.init()` fails outright on the new mappings:

  NumberFormatException: Value out of range. Value:"32768" Radix:10
      at java.lang.Short.parseShort
      at SimpleBlock.lambda$init$3(SimpleBlock.java:85)

Only the *state* id needs widening - the block registry id (max 1285 in 26.3)
and the item ids (max 1657) still fit and are left alone, as are the legacy
ids in BlockStorage17 and the 1.8 global palette.

BREAKING CHANGE: VirtualBlock#getModernID, #getID and #getBlockStateID now
return int instead of short. Downstream plugins (LimboAuth, LimboFilter, ...)
need to be recompiled against the new API.

+++ api/src/main/java/net/elytrium/limboapi/api/chunk/VirtualBlock.java
@@ -12,12 +12,12 @@ import net.elytrium.limboapi.api.material.WorldVersion;
 
 public interface VirtualBlock {
 
-  short getModernID();
+  int getModernID();
 
   String getModernStringID();
 
   @Deprecated
-  short getID(ProtocolVersion version);
+  int getID(ProtocolVersion version);
 
   short getBlockID(WorldVersion version);
 
@@ -27,7 +27,7 @@ public interface VirtualBlock {
 
   boolean isSupportedOn(WorldVersion version);
 
-  short getBlockStateID(ProtocolVersion version);
+  int getBlockStateID(ProtocolVersion version);
 
   boolean isSolid();
